Decent release #1
To say I am proud of what I have made so far is an understatement.
I learned how to host a Flask app on my local network which incorporates many aspects to make a functioning game. Tomorrow I will continue to refine this project until I feel satisfied to host it on a public platform.